Refactor app build for jupyter work
Summary:

Test Plan:
